Stripping down robot code to its generic basics for build team testing
Build teams wants a way to test a robot they're building up from
ponyo's parts, but with a different function on it.  This trips down
ponyo's code to its barest drivetrain basics, and then tacks on a
generic two-motor elevator and a generic single-motor intake as a
baseline for testing it out.